adminで使うapiはadmin/api
