+++ next/src/app/(admin)/admin/movie/page.tsx
@@ -3,7 +3,21 @@
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import MovieFilter from '@/app/_components/MovieFilter/MovieFilter';
-import { Movie } from '@/type/movie/movie';
+
+export type Movie = {
+  id: string;
+  title: string;
+  releaseDate: string;
+  duration: number;
+  genre: string;
+  description?: string;
+  director: string;
+  casts: string;
+  imageUrl?: string;
+  trailerUrl?: string;
+  showings: any[];
+  showingCount: number;
+};
 
 export default function MoviesPage() {
   const [movies, setMovies] = useState<Movie[]>([]);
